+++ docs/release_notes.rst
@@ -3,7 +3,7 @@ Release notes
 
 4.15.0
 
-*30 April 2026*
+*11 May 2026*
 
 - **Security**