+++ docs/release_notes.rst
@@ -77,7 +77,7 @@ more details regarding the lower-level changes required to accomplish it.*
 4.15.0
 ------
 
-*30 April 2026*
+*11 May 2026*
 
 - **Security**