+++ docs/release_notes.rst
@@ -1,6 +1,22 @@
 Release notes
 =============
 
+4.15.1
+------
+
+*19 May 2026*
+
+- **Security**
+
+  - Update dependencies to fix security vulnerabilities: mistune, urllib3, and
+    various UI dependencies (`PR#2608 <https://github.com/vantage6/vantage6/pull/2608>`_).
+
+- **Change**
+
+  - Update community algorithm store to use the v4 branch instead of the main branch.
+    This is to ensure v4 will still work when v5 is released.
+    (`PR#2607 <https://github.com/vantage6/vantage6/pull/2607>`_).
+
 4.15.0
 
 *11 May 2026*